# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ContextOS API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /adapters:
    get:
      operationId: listAdapters
      tags: [adapters]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Adapter"
    post:
      operationId: createAdapter
      tags: [adapters]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdapterInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Adapter"
  /adapters/{id}:
    get:
      operationId: getAdapter
      tags: [adapters]
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdapterDetail"
    patch:
      operationId: updateAdapter
      tags: [adapters]
      parameters:
        - $ref: "#/components/parameters/IdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdapterUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Adapter"
    delete:
      operationId: deleteAdapter
      tags: [adapters]
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "204":
          description: Deleted
  /adapters/{id}/discover:
    post:
      operationId: discoverAdapter
      tags: [adapters]
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdapterDetail"
  /adapters/{id}/test:
    post:
      operationId: testAdapter
      tags: [adapters]
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthTestResult"
  /capabilities:
    get:
      operationId: listCapabilities
      tags: [capabilities]
      parameters:
        - name: adapterId
          in: query
          required: false
          schema:
            type: string
        - name: type
          in: query
          required: false
          schema:
            type: string
        - name: riskTier
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Capability"
...
components:
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: string
  schemas:
...
    JsonObject:
      type: object
      additionalProperties: true
...
    Adapter:
      type: object
      required:
        [id, name, transport, protocolVersion, status, isGenerated, createdAt]
      properties:
        id: { type: string }
        name: { type: string }
        description: { type: ["string", "null"] }
        transport: { type: string }
        protocolVersion: { type: string }
        endpointUrl: { type: ["string", "null"] }
        sessionMode: { type: string }
        status: { type: string }
        linkedAccountId: { type: ["string", "null"] }
        isGenerated: { type: boolean }
        lastDiscoveredAt: { type: ["string", "null"], format: date-time }
        lastHealthAt: { type: ["string", "null"], format: date-time }
        capabilityCount: { type: ["integer", "null"] }
        createdAt: { type: string, format: date-time }
    AdapterDetail:
      allOf:
        - $ref: "#/components/schemas/Adapter"
        - type: object
          properties:
            capabilities:
              type: array
              items: { $ref: "#/components/schemas/Capability" }
            lastHealthResult: { $ref: "#/components/schemas/JsonObject" }
    AdapterInput:
      type: object
      required: [name, transport]
      properties:
        name: { type: string, minLength: 1 }
        description: { type: string }
        transport: { type: string }
        protocolVersion: { type: string }
        endpointUrl: { type: string }
        sessionMode: { type: string }
        linkedAccountId: { type: string }
    AdapterUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        description: { type: string }
        endpointUrl: { type: string }
        sessionMode: { type: string }
        status: { type: string }
    Capability:
      type: object
      required: [id, adapterId, type, name, riskTier, actionKind, createdAt]
      properties:
        id: { type: string }
        adapterId: { type: string }
        type: { type: string }
        name: { type: string }
        description: { type: ["string", "null"] }
        riskTier: { type: string }
        actionKind: { type: string }
        humanReviewRequired: { type: boolean }
        inputSchema: { $ref: "#/components/schemas/JsonObject" }
        outputSchema: { $ref: "#/components/schemas/JsonObject" }
        createdAt: { type: string, format: date-time }
    HealthTestResult:
      type: object
      required: [ok, latencyMs, message]
      properties:
        ok: { type: boolean }
        latencyMs: { type: integer }
        message: { type: string }
        details: { $ref: "#/components/schemas/JsonObject" }
...